Fixed tests for non Windows
